Improve moving when the zoom is too much
Now, when you zoom, the movement of the mouse in the 'virtual' screen is
proportional to the real screen, so that when you zoom a lot, the mouse
can reach all the corners of the real screen, because the pixmap will move
with it.

Before, the movement of the 'virtual' screen was given by a constant
multiplied by the movement of the mouse. This caused to not being able
to reach the corners of the real screen when the zoom was big enough to
trigger this error.