Fix a bug when it loses focus
If the app loses focus (like moving the cursor to another screen and
moving it back in, or calling a shortcut that makes focus in another window,
etc.), it couldn't read the keys pressed, so you were not able to
exit. This commit fixes this, by requesting focus when the mouse moves
inside ZoomMe if it's not active (its window).