Fix a bug that caused RAM to overflow
Before, the frames were sent directly to FFmpeg, but that caused the RAM
to overflow very quickly because ZoomMe was sending raw images 60 times
a second (so... 4 bytes per pixel * your resolution * 60 times a second
= A LOT of Megabytes per second).

Now, the frames are saved in a temporal file in the disk, the images are
compressed and with less Frames per second (16). When it finishes
recording, `cat` reads the file, piping it to `ffmpeg`, causing
considerables less memory usage.